+++ XuanYu.Editor.UI/Viewport/Vulkan/VulkanViewport.axaml.cs
@@ -1,4 +1,5 @@
 using Avalonia.Controls;
+using Avalonia.VisualTree;
 
 namespace XuanYu.Editor.UI;
 
@@ -7,6 +8,7 @@ public partial class VulkanViewport : UserControl
     public VulkanViewport()
     {
         InitializeComponent();
+        this.GetVisualDescendants().OfType<VulkanNativeHost>().Single().RendererReady += (_, _) => HideFallback();
         AttachedToVisualTree += (_, _) => SetFallback("Vulkan 正在初始化...");
     }